Configure sms to select chain and set coupon selection to random

+++ server/sms/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 
 class SMS(models.Model):
-    number = models.PositiveIntegerField(unique=True, validators=[MaxValueValidator(9999999999)])
+    number = models.BigIntegerField(unique=True, validators=[MaxValueValidator(9999999999)])
     message_body = models.CharField(max_length=140)
 
     def __unicode__(self):
